requirements.txt
# Web Development and Static Site Tools
beautifulsoup4>=4.12.0    # HTML/XML parsing
requests>=2.31.0          # HTTP library
html5lib>=1.1             # HTML5 parser
cssutils>=2.9.0           # CSS parsing and manipulation
Pillow>=10.0.0            # Image processing
markdown>=3.5.0           # Markdown processing
jinja2>=3.1.0             # Template engine
pyyaml>=6.0               # YAML processing

# Code Quality and Formatting
black>=23.0.0             # Code formatter
flake8>=6.0.0             # Linting
pylint>=2.17.0            # Code analysis
mypy>=1.5.0               # Type checking

# Development Tools
ipython>=8.15.0           # Enhanced Python shell
jupyter>=1.0.0            # Notebook environment
pre-commit>=3.4.0         # Git hooks

# Testing
pytest>=7.4.0            # Testing framework
pytest-cov>=4.1.0        # Coverage plugin

# Static Site Generation (optional)
pelican>=4.8.0            # Static site generator
mkdocs>=1.5.0             # Documentation generator

# Automation and Deployment
fabric>=3.2.0             # Deployment automation
invoke>=2.2.0             # Task runner

# Data and API Tools
python-dotenv>=1.0.0      # Environment variables
click>=8.1.0              # CLI creation
rich>=13.5.0              # Rich text and beautiful formatting
